+++ src/plone/restapi/serializer/local_roles.py
@@ -17,10 +17,11 @@ class SerializeLocalRolesToJson(object):
         self.context = context
         self.request = request
 
-    def __call__(self):
+    def __call__(self, search=None):
+        self.request.form['search_term'] = search
         sharing_view = getMultiAdapter((self.context, self.request),
                                        name='sharing')
-        local_roles = sharing_view.existing_role_settings()
+        local_roles = sharing_view.role_settings()
         available_roles = [r['id'] for r in sharing_view.roles()]
 
         blocked_roles = getattr(