+++ src/plone/restapi/tests/test_content_local_roles.py
@@ -32,6 +32,13 @@ class TestFolderCreate(unittest.TestCase):
         )
         wftool = getToolByName(self.portal, 'portal_workflow')
         wftool.doActionFor(self.portal.folder1, 'publish')
+
+        self.portal.folder1.invokeFactory(
+            'Document',
+            id='doc1',
+            title='My Document'
+        )
+
         transaction.commit()
 
     def _get_ac_local_roles_block(self, obj):
@@ -39,6 +46,73 @@ class TestFolderCreate(unittest.TestCase):
             getattr(aq_base(self.portal.folder1),
                     '__ac_local_roles_block__',
                     False))
+
+    def test_sharing_search(self):
+        '''A request to @sharing should support the search parameter. '''
+        response = requests.get(
+            self.portal.folder1.absolute_url() + '/@sharing',
+            headers={'Accept': 'application/json'},
+            auth=(SITE_OWNER_NAME, SITE_OWNER_PASSWORD),
+        )
+        non_search_entries = response.json()['entries']
+
+        response = requests.get(
+            self.portal.folder1.absolute_url() + '/@sharing?search=admin',
+            headers={'Accept': 'application/json'},
+            auth=(SITE_OWNER_NAME, SITE_OWNER_PASSWORD),
+        )
+        search_entries = response.json()['entries']
+
+        # Did we find anything?
+        self.assertNotEqual(len(non_search_entries), len(search_entries))
+
+    def test_sharing_search_roundtrip(self):
+        '''Search for a user and use save roles
+        '''
+        # Make sure we don't already have admin
+        response = requests.get(
+            self.portal.folder1.absolute_url() + '/@sharing',
+            headers={'Accept': 'application/json'},
+            auth=(SITE_OWNER_NAME, SITE_OWNER_PASSWORD),
+        )
+        self.assertNotIn(
+            'admin', [x['id'] for x in response.json()['entries']]
+        )
+
+        # Now find admin and set roles
+        response = requests.get(
+            self.portal.folder1.absolute_url() + '/@sharing?search=admin',
+            headers={'Accept': 'application/json'},
+            auth=(SITE_OWNER_NAME, SITE_OWNER_PASSWORD),
+        )
+        roles = [x for x in response.json()['entries'] if x['id'] == 'admin']
+        roles = roles[0]['roles']
+
+        new_roles = dict([(key, not val) for key, val in roles.items()])
+        payload = {'entries': [{'id': 'admin', 'roles': new_roles}]}
+
+        response = requests.post(
+            self.portal.folder1.absolute_url() + '/@sharing',
+            headers={'Accept': 'application/json'},
+            auth=(SITE_OWNER_NAME, SITE_OWNER_PASSWORD),
+            json=payload,
+        )
+        self.assertEqual(204, response.status_code)
+
+        # Now we should have admin in @sharing
+        response = requests.get(
+            self.portal.folder1.absolute_url() + '/@sharing',
+            headers={'Accept': 'application/json'},
+            auth=(SITE_OWNER_NAME, SITE_OWNER_PASSWORD),
+        )
+        self.assertIn(
+            'admin', [x['id'] for x in response.json()['entries']]
+        )
+
+        # with the same roles as set
+        roles = [x for x in response.json()['entries'] if x['id'] == 'admin']
+        roles = roles[0]['roles']
+        self.assertEqual(new_roles, roles)
 
     def test_sharing_requires_delegate_roles_permission(self):
         '''A response for an object without any roles assigned'''
@@ -71,7 +145,7 @@ class TestFolderCreate(unittest.TestCase):
                             u'Reviewer': False},
                  u'title': u'Logged-in users',
                  u'type': u'group'}],
-             u'inherit': False}
+             u'inherit': True}
         )
 
     def test_get_local_roles_with_user(self):
@@ -110,7 +184,7 @@ class TestFolderCreate(unittest.TestCase):
                                u'Reviewer': True},
                     u'title': u'test-user',
                     u'type': u'user'}],
-             u'inherit': False}
+             u'inherit': True}
         )
 
     def test_set_local_roles_for_user(self):
@@ -185,6 +259,8 @@ class TestFolderCreate(unittest.TestCase):
         )
 
     def test_get_local_roles_inherit_roles(self):
+        # __ac_local_roles_block__ specifies to block inheritance:
+        # https://docs.plone.org/develop/plone/security/local_roles.html
         self.portal.folder1.__ac_local_roles_block__ = True
         transaction.commit()
 
@@ -195,7 +271,7 @@ class TestFolderCreate(unittest.TestCase):
         )
 
         self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.json()['inherit'], True)
+        self.assertEqual(response.json()['inherit'], False)
 
     def test_set_local_roles_inherit(self):
         self.assertEqual(self._get_ac_local_roles_block(self.portal.folder1),
@@ -245,6 +321,46 @@ class TestFolderCreate(unittest.TestCase):
         response = response.json()
         self.assertIn('available_roles', response)
         self.assertIn('Reader', response['available_roles'])
+
+    def test_inherited_global(self):
+        api.user.grant_roles(username=TEST_USER_ID, roles=['Reviewer'])
+        api.user.grant_roles(
+            username=TEST_USER_ID, obj=self.portal.folder1, roles=['Editor']
+        )
+        transaction.commit()
+
+        response = requests.get(
+            self.portal.folder1.doc1.absolute_url() + '/@sharing',
+            headers={'Accept': 'application/json'},
+            auth=(SITE_OWNER_NAME, SITE_OWNER_PASSWORD),
+        )
+
+        response = response.json()
+        # find our entry
+        entry = [x for x in response['entries'] if x['id'] == TEST_USER_ID][0]
+
+        self.assertEqual('global', entry['roles']['Reviewer'])
+        self.assertEqual('acquired', entry['roles']['Editor'])
+
+    def test_inherited_global_via_search(self):
+        api.user.create(email='jos@henken.local', username='jos')
+        api.user.grant_roles(username='jos', roles=['Reviewer'])
+        api.user.grant_roles(
+            username='jos', roles=['Editor'], obj=self.portal.folder1
+        )
+        transaction.commit()
+
+        response = requests.get(
+            self.portal.folder1.doc1.absolute_url() + '/@sharing?search=jos',
+            headers={'Accept': 'application/json'},
+            auth=(SITE_OWNER_NAME, SITE_OWNER_PASSWORD),
+        )
+        response = response.json()
+        # find our entry
+        entry = [x for x in response['entries'] if x['id'] == 'jos'][0]
+
+        self.assertEqual('global', entry['roles']['Reviewer'])
+        self.assertEqual('acquired', entry['roles']['Editor'])
 
     def test_no_serializer_available_returns_501(self):
         # This test unregisters the local_roles adapter. The testrunner can