+++ CHANGES.rst
@@ -59,6 +59,10 @@ Bugfixes:
 
 New Features:
 
+- Add support for a 'search' parameter to @sharing. This returns additional
+  principals in 'entries', also flagging the acquired and inherited fields.
+  [jaroel]
+
 - Add support for setting/modifying 'layout' on DX and AT content endpoints.
   [jaroel]